Organize render functions better
