+++ packages/core/src/cells/image-cell.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/display-name */
 import * as React from "react";
 import { ImageOverlayEditor } from "../internal/data-grid-overlay-editor/private/image-overlay-editor.js";
-import { roundedRect } from "../internal/data-grid/data-grid-lib.js";
+import { roundedRect } from "../internal/data-grid/render/data-grid-lib.js";
 import { GridCellKind, type BaseGridCell, type ImageCell } from "../internal/data-grid/data-grid-types.js";
 import type { BaseDrawArgs, InternalCellRenderer } from "./cell-types.js";
