+++ packages/core/src/cells/loading-cell.tsx
@@ -1,5 +1,5 @@
 import { withAlpha } from "../internal/data-grid/color-parser.js";
-import { roundedRect } from "../internal/data-grid/data-grid-lib.js";
+import { roundedRect } from "../internal/data-grid/render/data-grid-lib.js";
 import { GridCellKind, type LoadingCell } from "../internal/data-grid/data-grid-types.js";
 import type { InternalCellRenderer } from "./cell-types.js";
