+++ packages/core/src/cells/marker-cell.tsx
@@ -1,6 +1,6 @@
-import { getMiddleCenterBias } from "../internal/data-grid/data-grid-lib.js";
+import { getMiddleCenterBias } from "../internal/data-grid/render/data-grid-lib.js";
 import { InnerGridCellKind, type MarkerCell } from "../internal/data-grid/data-grid-types.js";
-import { drawCheckbox } from "../internal/data-grid/draw-checkbox.js";
+import { drawCheckbox } from "../internal/data-grid/render/draw-checkbox.js";
 import type { BaseDrawArgs, InternalCellRenderer, PrepResult } from "./cell-types.js";
 
 export const markerCellRenderer: InternalCellRenderer<MarkerCell> = {
