+++ packages/core/src/cells/number-cell.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/display-name */
 import * as React from "react";
-import { drawTextCell, prepTextCell } from "../internal/data-grid/data-grid-lib.js";
+import { drawTextCell, prepTextCell } from "../internal/data-grid/render/data-grid-lib.js";
 import { GridCellKind, type NumberCell } from "../internal/data-grid/data-grid-types.js";
 import type { InternalCellRenderer } from "./cell-types.js";
