+++ packages/core/src/docs/examples/row-hover.stories.tsx
@@ -7,7 +7,7 @@ import {
     defaultProps,
     useAllMockedKinds,
 } from "../../data-editor/stories/utils.js";
-import type { GetRowThemeCallback } from "../../internal/data-grid/data-grid-render.js";
+import type { GetRowThemeCallback } from "../../internal/data-grid/render/data-grid-render.cells.js";
 import { SimpleThemeWrapper } from "../../stories/story-utils.js";
 import type { GridMouseEventArgs } from "../../internal/data-grid/event-args.js";
