+++ packages/core/src/index.ts
@@ -5,7 +5,7 @@ export type { SpriteProps } from "./common/utils.js";
 export type { Theme } from "./common/styles.js";
 export type { CustomRenderer, BaseDrawArgs, DrawArgs } from "./cells/cell-types.js";
 export type { SelectionBlending } from "./internal/data-grid/use-selection-behavior.js";
-export type { GetRowThemeCallback, Highlight } from "./internal/data-grid/data-grid-render.js";
+export type { GetRowThemeCallback, Highlight } from "./internal/data-grid/render/data-grid-render.cells.js";
 export type { ImageWindowLoader } from "./internal/data-grid/image-window-loader-interface.js";
 export * from "./internal/data-grid/data-grid-types.js";
 export type {
@@ -36,7 +36,7 @@ export {
     roundedPoly,
     roundedRect,
     drawTextCellExternal as drawTextCell,
-} from "./internal/data-grid/data-grid-lib.js";
+} from "./internal/data-grid/render/data-grid-lib.js";
 export { CellSet } from "./internal/data-grid/cell-set.js";
 export { getDataEditorTheme as getDefaultTheme, useTheme } from "./common/styles.js";
 export { useColumnSizer } from "./data-editor/use-column-sizer.js";
